#210 moving get refresh token to Token controller
