+++ API/ControllerLogic/TokenControllerLogic.cs
@@ -1,7 +1,10 @@
-﻿using DataLayer.Mongo.Entities;
+﻿using Common;
+using DataLayer.Mongo.Entities;
 using DataLayer.Mongo.Repositories;
 using Encryption;
 using Microsoft.AspNetCore.Mvc;
+using System.IdentityModel.Tokens.Jwt;
+using System.Reflection;
 using System.Security.Cryptography;
 
 namespace API.ControllerLogic
@@ -9,36 +12,98 @@ namespace API.ControllerLogic
     public class TokenControllerLogic : ITokenControllerLogic
     {
         private readonly IUserRepository _userRepository;
-        public TokenControllerLogic(IUserRepository userRepository)
+        private readonly IMethodBenchmarkRepository _methodBenchmarkRepository;
+        private readonly IEASExceptionRepository _exceptionRepository;
+        public TokenControllerLogic(
+            IUserRepository userRepository,
+            IMethodBenchmarkRepository methodBenchmarkRepository,
+            IEASExceptionRepository exceptionRepository)
         {
             this._userRepository = userRepository;
+            this._methodBenchmarkRepository = methodBenchmarkRepository;
+            this._exceptionRepository = exceptionRepository;
         }
         #region GetToken
         public async Task<IActionResult> GetToken(HttpContext httpContext)
         {
+            BenchmarkMethodLogger logger = new BenchmarkMethodLogger(httpContext);
             IActionResult result = null;
-            string apiKey = httpContext.Request.Headers["ApiKey"].ToString();
-            if (string.IsNullOrEmpty(apiKey))
+            try
             {
-                result = new UnauthorizedObjectResult(new { error = "You did not set an ApiKey" });
+                string apiKey = httpContext.Request.Headers["ApiKey"].ToString();
+                if (string.IsNullOrEmpty(apiKey))
+                {
+                    result = new UnauthorizedObjectResult(new { error = "You did not set an ApiKey" });
+                }
+                User activeUser = await this._userRepository.GetUserByApiKey(apiKey);
+                if (activeUser == null)
+                {
+                    result = new UnauthorizedObjectResult(new { error = "You entered an invalid ApiKey" });
+                }
+                else if (activeUser.IsActive == false)
+                {
+                    result = new UnauthorizedObjectResult(new { error = "This user account is no longer active" });
+                }
+                else
+                {
+                    RSACryptoServiceProvider RSAalg = new RSACryptoServiceProvider(4096);
+                    RSAParameters rsaParams = RSAalg.ExportParameters(true);
+                    string publicKey = RSAalg.ToXmlString(false);
+                    string token = new JWT().GenerateSecurityToken(activeUser.Id, rsaParams, publicKey, activeUser.IsAdmin);
+                    result = new OkObjectResult(new { token = token });
+                }
+                return result;
+
             }
-            User activeUser = await this._userRepository.GetUserByApiKey(apiKey);
-            if (activeUser == null)
+            catch (Exception ex)
             {
-                result = new UnauthorizedObjectResult(new { error = "You entered an invalid ApiKey"});
+                result = new BadRequestObjectResult(new { error = "Something went wrong on our end"});
+                await this._exceptionRepository.InsertException(ex.ToString(), MethodBase.GetCurrentMethod().Name);
             }
-            else if (activeUser.IsActive == false)
+            return result;
+        }
+        #endregion
+
+        #region GetRefreshToken
+        public async Task<IActionResult> GetRefreshToken(HttpContext context)
+        {
+            BenchmarkMethodLogger logger = new BenchmarkMethodLogger(context);
+            IActionResult result = null;
+            try
             {
-                result = new UnauthorizedObjectResult(new { error = "This user account is no longer active" });
+                // get current token
+                string token = context.Request.Headers["Authorization"].FirstOrDefault()?.Split(" ").Last();
+                if (!string.IsNullOrEmpty(token))
+                {
+                    var handler = new JwtSecurityTokenHandler().ReadJwtToken(token);
+                    JWT jwtWrapper = new JWT();
+                    string publicKey = handler.Claims.First(x => x.Type == "public-key").Value;
+                    RSACryptoServiceProvider rsaProvdier = new RSACryptoServiceProvider(4096);
+                    rsaProvdier.FromXmlString(publicKey);
+                    RSAParameters parameters = rsaProvdier.ExportParameters(false);
+                    if (!await jwtWrapper.ValidateSecurityToken(token, parameters))
+                    {
+                        RSAProviderWrapper rsa4096 = new RSAProviderWrapper(4096);
+                        string userId = jwtWrapper.GetUserIdFromToken(token);
+                        bool isAdmin = bool.Parse(handler.Claims.First(x => x.Type == "IsAdmin").Value);
+                        string newToken = new JWT().GenerateSecurityToken(userId, rsa4096.rsaParams, rsa4096.publicKey, isAdmin);
+                        result = new OkObjectResult(new { token = newToken });
+                    }
+                    else
+                    {
+                        // if token is still valid just send the same token back.
+                        result = new OkObjectResult(new { token = token });
+                    }
+                }
             }
-            else
+            catch (Exception ex)
             {
-                RSACryptoServiceProvider RSAalg = new RSACryptoServiceProvider(4096);
-                RSAParameters rsaParams = RSAalg.ExportParameters(true);
-                string publicKey = RSAalg.ToXmlString(false);
-                string token = new JWT().GenerateSecurityToken(activeUser.Id, rsaParams, publicKey, activeUser.IsAdmin);
-                result = new OkObjectResult(new { token = token });
+                await this._exceptionRepository.InsertException(ex.ToString(), MethodBase.GetCurrentMethod().Name);
+                // TODO: give error message
+                result = new BadRequestObjectResult(new { });
             }
+            logger.EndExecution();
+            await this._methodBenchmarkRepository.InsertBenchmark(logger);
             return result;
         }
         #endregion
